Script-Style Process Types
Previously, whenever a Procfile was parsed and turned into collection of
process types, the arguments were always shell-parsed.  It turns out that
Procfiles commonly have inline environment variables (FOO=bar ./my-app), and
joined commands (./my-app-1 && ./my-app-2).  None of these things shell-parse
well.

This change update the behavior so that the command is shell-parsed if it's
going onto `tiny` (a shell-less stack), but will leave it as "script-style"
command for all stacks with a shell.

Signed-off-by: Ben Hale <bhale@vmware.com>